+++ CHANGELOG.md
@@ -7,6 +7,7 @@ v1.2.x
 - More efficiency in internal send methods, thank you @marcussacana
 - Removal of Newtonsoft.Json
 - Dependency update
+- Case insensitive dictionaries
 
 ## Previous Versions